integration tests clusters

See merge request cse2115/2022-2023/SEM09b!47

+++ jobs-microservice/src/main/java/nl/tudelft/sem/template/example/domain/JobService.java
@@ -3,6 +3,7 @@ package nl.tudelft.sem.template.example.domain;
 import commons.Faculty;
 import commons.Job;
 import commons.NetId;
+import commons.Role;
 import commons.RoleValue;
 import commons.ScheduleJob;
 import commons.Status;
@@ -95,7 +96,7 @@ public class JobService {
      * @throws Exception if the resources of NetId are invalid
      */
 
-    public Job createJob(NetId netId, NetId authNetId, Faculty faculty, int cpuUsage, int gpuUsage,
+    public Job createJob(NetId netId, NetId authNetId, Faculty faculty, String desc, int cpuUsage, int gpuUsage,
                          int memoryUsage, RoleValue role, LocalDate preferredDate) throws Exception {
         if (cpuUsage < 0 || gpuUsage < 0 || memoryUsage < 0) {
             throw new InvalidResourcesException(Math.min(cpuUsage, Math.min(gpuUsage, memoryUsage)));
@@ -110,12 +111,11 @@ public class JobService {
         if (!netId.toString().equals(authNetId.toString())) {
             throw new InvalidNetIdException(netId.toString());
         }
-        if (!role.equals(RoleValue.EMPLOYEE)) {
-            System.out.println(role);
+        if (role != RoleValue.EMPLOYEE) {
             throw new BadCredentialsException(role.toString());
         }
 
-        Job newJob = new Job(netId, faculty, cpuUsage, gpuUsage, memoryUsage, preferredDate);
+        Job newJob = new Job(netId, faculty, desc, cpuUsage, gpuUsage, memoryUsage, preferredDate);
         jobRepository.save(newJob);
 
         return newJob;
@@ -141,8 +141,7 @@ public class JobService {
         if (!job.getNetId().toString().equals(authNetId.toString())) {
             throw new InvalidNetIdException(job.getNetId().toString());
         }
-        if (!role.equals(RoleValue.EMPLOYEE)) {
-            System.out.println(role);
+        if (role != RoleValue.EMPLOYEE) {
             throw new BadCredentialsException(role.toString());
         }
 
@@ -157,11 +156,17 @@ public class JobService {
      * @param id the unique id of the Job.
      * @throws Exception if there is no Job with the provided id.
      */
-    public void deleteJob(long id) throws Exception {
+    public void deleteJob(String netId, RoleValue role, long id) throws Exception {
         if (!jobRepository.existsById(id)) {
             throw new InvalidIdException(id);
         }
-        jobRepository.deleteById(id);
+
+        if (role == RoleValue.ADMIN || ((role == RoleValue.EMPLOYEE)
+            && netId.equals(jobRepository.getOne(id).getNetId().toString()))) {
+            jobRepository.deleteById(id);
+        } else {
+            throw new BadCredentialsException("Not admin or not your job");
+        }
     }
 
     /**
@@ -216,15 +221,15 @@ public class JobService {
      * @return a list of Job entities containing all jobs in the database.
      * @throws Exception if the NetId is invalid or the creator of the request does not have the admin role.
      */
-    public List<Job> getAllJobs(NetId netId, NetId authNetId, String role) throws Exception {
+    public List<Job> getAllJobs(NetId netId, NetId authNetId, RoleValue role) throws Exception {
         if (netId == null) {
             throw new InvalidNetIdException(nullValue);
         }
         if (!netId.toString().equals(authNetId.toString())) {
             throw new InvalidNetIdException(netId.toString());
         }
-        if (!role.equals("admin")) {
-            throw new BadCredentialsException(role);
+        if (role != RoleValue.ADMIN) {
+            throw new BadCredentialsException(role.toString());
         }
         return jobRepository.findAll();
     }
@@ -239,15 +244,15 @@ public class JobService {
      * @return a list of Job entities containing all "ACCEPTED" jobs in the database.
      * @throws Exception if the NetId is invalid or the creator of the request does not have the admin role.
      */
-    public List<Job> getAllScheduledJobs(NetId netId, NetId authNetId, String role) throws Exception {
+    public List<Job> getAllScheduledJobs(NetId netId, NetId authNetId, RoleValue role) throws Exception {
         if (netId == null) {
             throw new InvalidNetIdException(nullValue);
         }
         if (!netId.toString().equals(authNetId.toString())) {
             throw new InvalidNetIdException(netId.toString());
         }
-        if (!role.equals("admin")) {
-            throw new BadCredentialsException(role);
+        if (role != RoleValue.ADMIN) {
+            throw new BadCredentialsException(role.toString());
         }
         return jobRepository.findAll().stream().filter(j -> j.getStatus() == Status.ACCEPTED).collect(Collectors.toList());
     }
